Use control sockets to read configuration
The exporter now consumes the control sockets of kea instances and
pulls the configuration from it. This allows the use of more complex
configurations, including those using includes. This allows us to
drop the dependency on hjson.

Config reloading through inotify is not working, so config reloads
now happen before every stats pull.

Updated the README with new usage, features and updated limitations.